# HW/2/openapi.yaml
openapi: 3.0.3
info:
  title: Hotel Booking API
  description: REST API для системы бронирования отелей (Вариант 13)
  version: "1.0.0"

servers:
  - url: http://localhost:8000
    description: Local dev server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRegister:
      type: object
      required: [login, password, first_name, last_name, email]
      properties:
        login:
          type: string
          minLength: 3
          maxLength: 50
          example: ivan_petrov
        password:
          type: string
          minLength: 4
          example: mypassword123
        first_name:
          type: string
          example: Ivan
        last_name:
          type: string
          example: Petrov
        email:
          type: string
          example: ivan@example.com

    UserLogin:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          example: ivan_petrov
        password:
          type: string
          example: mypassword123

    Token:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer

    UserOut:
      type: object
      properties:
        id:
          type: string
        login:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string

    HotelCreate:
      type: object
      required: [name, city, address, stars, rooms_total, price_per_night]
      properties:
        name:
          type: string
          example: Grand Hotel Moscow
        city:
          type: string
          example: Moscow
        address:
          type: string
          example: Tverskaya 1
        stars:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        rooms_total:
          type: integer
          minimum: 1
          example: 100
        price_per_night:
          type: number
          example: 5000.0

    HotelOut:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        address:
          type: string
        stars:
          type: integer
        rooms_total:
          type: integer
        price_per_night:
          type: number

    BookingCreate:
      type: object
      required: [hotel_id, check_in, check_out]
      properties:
        hotel_id:
          type: string
          example: abc123def456
        check_in:
          type: string
          format: date
          example: "2026-06-01"
        check_out:
          type: string
          format: date
          example: "2026-06-05"

    BookingOut:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        hotel_id:
          type: string
        hotel_name:
          type: string
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
        total_price:
          type: number
        status:
          type: string
          enum: [confirmed, cancelled]
        created_at:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        detail:
          type: string

paths:
  /:
    get:
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: OK

  /api/auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
        "409":
          description: Login already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/auth/login:
    post:
      tags: [auth]
      summary: Login and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/users/search:
    get:
      tags: [users]
      summary: Search users by login or name mask
      security:
        - BearerAuth: []
      parameters:
        - name: login
          in: query
          schema:
            type: string
          description: Exact login match
        - name: name
          in: query
          schema:
            type: string
          description: Case-insensitive substring in first_name + last_name
      responses:
        "200":
          description: List of matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserOut"
        "400":
          description: No search params provided
        "401":
          description: Unauthorized

  /api/users/{user_id}:
    get:
      tags: [users]
      summary: Get user by ID
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
        "404":
          description: User not found

  /api/hotels:
    get:
      tags: [hotels]
      summary: Get list of all hotels
      responses:
        "200":
          description: List of hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HotelOut"
    post:
      tags: [hotels]
      summary: Create a new hotel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HotelCreate"
      responses:
        "201":
          description: Hotel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelOut"
        "401":
          description: Unauthorized

  /api/hotels/search:
    get:
      tags: [hotels]
      summary: Search hotels by city
      parameters:
        - name: city
          in: query
          required: true
          schema:
            type: string
          description: City name (case-insensitive substring)
      responses:
        "200":
          description: Matching hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HotelOut"

  /api/hotels/{hotel_id}:
    get:
      tags: [hotels]
      summary: Get hotel by ID
      parameters:
        - name: hotel_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hotel details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelOut"
        "404":
          description: Hotel not found

  /api/bookings:
    post:
      tags: [bookings]
      summary: Create a new booking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingCreate"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingOut"
        "400":
          description: Invalid dates
        "401":
          description: Unauthorized
        "404":
          description: Hotel not found

  /api/bookings/my:
    get:
      tags: [bookings]
      summary: Get current user's bookings
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingOut"
        "401":
          description: Unauthorized

  /api/bookings/{booking_id}:
    get:
      tags: [bookings]
      summary: Get booking by ID
      security:
        - BearerAuth: []
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingOut"
        "401":
          description: Unauthorized
        "403":
          description: Access denied
        "404":
          description: Booking not found
    delete:
      tags: [bookings]
      summary: Cancel a booking
      security:
        - BearerAuth: []
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking cancelled
        "400":
          description: Already cancelled
        "401":
          description: Unauthorized
        "403":
          description: Access denied
        "404":
          description: Booking not found